* lily/paper-column.cc (before_line_breaking): bug fix.
* lily/volta-engraver.cc (process_music): do something sensible if
no staffs found.

* scripts/lilypond-book.py (make_pixmap): opps. Do png conversion
only once.

* Documentation/user/refman.itely (Global layout): made section on
global layout, added spacing documentation.

* lily/staff-symbol.cc (brew_molecule): change ragged right for
staff symbols. See comments.

+++ scripts/lilypond-book.py
@@ -1210,11 +1210,13 @@ def make_pixmap (name):
 	cmd = r'''gs -g%dx%d -sDEVICE=pnggray  -dTextAlphaBits=4 -dGraphicsAlphaBits=4  -q -sOutputFile=- -r%d -dNOPAUSE %s %s -c quit  > %s'''
 	
 	cmd = cmd % (x, y, res, name + '.trans.eps', name + '.eps',name + '.png')
-	quiet_system (cmd, 'gs')
-
+	status = 0
 	try:
-		status = system (cmd)
+		status = quiet_system (cmd, 'gs')
 	except:
+		status = -1
+
+	if status:
 		os.unlink (name + '.png')
 		error ("Removing output file")
